fixing CI

+++ src/libAtomVM/context.h
@@ -125,7 +125,7 @@ struct Context
     // Ports support
     native_handler_f native_handler;
 
-    uint64_t reductions;
+    int reductions;
 
     unsigned int leader : 1;
     unsigned int has_min_heap_size : 1;
